+++ lib/WebServer/DirIndex/HTML.pm
@@ -2,10 +2,20 @@ use strict;
 use warnings;
 use Feature::Compat::Class;
 
-class WebServer::DirIndex::HTML v0.0.2 {
+class WebServer::DirIndex::HTML v0.0.3 {
 
   field $file_html :reader = <<'FILE';
   <tr>
+    <td class='name'><a href='%s'>%s</a></td>
+    <td class='size'>%s</td>
+    <td class='type'>%s</td>
+    <td class='mtime'>%s</td>
+  </tr>
+FILE
+
+  field $file_html_icons :reader = <<'FILE';
+  <tr>
+    <td class='icon'><i class='%s'></i></td>
     <td class='name'><a href='%s'>%s</a></td>
     <td class='size'>%s</td>
     <td class='type'>%s</td>
@@ -28,6 +38,38 @@ FILE
     <table>
       <thead>
         <tr>
+          <th class='name'>Name</th>
+          <th class='size'>Size</th>
+          <th class='type'>Type</th>
+          <th class='mtime'>Last Modified</th>
+        </tr>
+      </thead>
+      <tbody>
+%s
+      </tbody>
+    </table>
+    <hr />
+  </body>
+</html>
+DIR
+
+  field $dir_html_icons :reader = <<'DIR';
+<html>
+  <head>
+    <title>%s</title>
+    <meta http-equiv="content-type" content="text/html; charset=utf-8" />
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css" />
+    <style type='text/css'>
+%s
+    </style>
+  </head>
+  <body>
+    <h1>%s</h1>
+    <hr />
+    <table>
+      <thead>
+        <tr>
+          <th class='icon'></th>
           <th class='name'>Name</th>
           <th class='size'>Size</th>
           <th class='type'>Type</th>
@@ -71,12 +113,23 @@ index page. The actual rendering is performed by L<WebServer::DirIndex>.
 
 =item file_html
 
-Returns a C<sprintf> format string used to render a single file row.
+Returns a C<sprintf> format string used to render a single file row
+(without an icon column). Contains 5 C<%s> placeholders.
+
+=item file_html_icons
+
+Returns a C<sprintf> format string used to render a single file row
+with a Font Awesome icon in the first column. Contains 6 C<%s> placeholders.
 
 =item dir_html
 
 Returns a C<sprintf> format string used to render the full directory
-index page.
+index page (without icons).
+
+=item dir_html_icons
+
+Returns a C<sprintf> format string used to render the full directory
+index page with a Font Awesome CDN link and an icon column header.
 
 =back
 
@@ -86,7 +139,7 @@ You can subclass this module to provide custom HTML templates. Override
 C<file_html> and/or C<dir_html> in your subclass by declaring new fields
 with the C<:reader> attribute.
 
-Both templates are C<sprintf> format strings. The placeholders (C<%s>) are
+All templates are C<sprintf> format strings. The placeholders (C<%s>) are
 positional and must be preserved in the correct order:
 
 =over 4
@@ -95,7 +148,15 @@ positional and must be preserved in the correct order:
 
 C<url>, C<name>, C<size>, C<mime_type>, C<mtime>.
 
+=item file_html_icons placeholders (6 total, in order)
+
+C<icon_class>, C<url>, C<name>, C<size>, C<mime_type>, C<mtime>.
+
 =item dir_html placeholders (4 total, in order)
+
+Page C<title>, inline C<css>, page C<heading>, C<file rows>.
+
+=item dir_html_icons placeholders (4 total, in order)
 
 Page C<title>, inline C<css>, page C<heading>, C<file rows>.
 
@@ -103,6 +164,10 @@ Page C<title>, inline C<css>, page C<heading>, C<file rows>.
 
 Pass your subclass name as the C<html_class> parameter when constructing
 L<WebServer::DirIndex> or L<WebServer::DirIndex::File>.
+
+When C<icons> is enabled in L<WebServer::DirIndex>, the C<file_html_icons>
+and C<dir_html_icons> templates are used instead of C<file_html> and
+C<dir_html>, provided the html class supports them (detected via C<can>).
 
 =head1 AUTHOR